+++ python/src/tel_grammar/antlr/TelParser.py
@@ -11,34 +11,36 @@ else:
 
 def serializedATN():
     with StringIO() as buf:
-        buf.write("\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\3\35")
-        buf.write("N\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\3\2")
+        buf.write("\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964\3\37")
+        buf.write("T\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\3\2")
         buf.write("\3\2\3\2\5\2\22\n\2\3\2\3\2\7\2\26\n\2\f\2\16\2\31\13")
         buf.write("\2\3\2\3\2\3\3\3\3\3\3\5\3 \n\3\3\3\3\3\5\3$\n\3\3\4\5")
         buf.write("\4\'\n\4\3\4\3\4\3\5\3\5\3\5\3\6\3\6\3\6\3\6\5\6\62\n")
-        buf.write("\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\7\6=\n\6\f\6\16")
-        buf.write("\6@\13\6\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\3\7\5\7L")
-        buf.write("\n\7\3\7\2\3\n\b\2\4\6\b\n\f\2\7\3\2\32\33\3\2\30\31\3")
-        buf.write("\2\20\27\3\2\3\4\3\2\5\6\2V\2\16\3\2\2\2\4\34\3\2\2\2")
-        buf.write("\6&\3\2\2\2\b*\3\2\2\2\n\61\3\2\2\2\fK\3\2\2\2\16\17\7")
-        buf.write("\b\2\2\17\21\7\13\2\2\20\22\5\n\6\2\21\20\3\2\2\2\21\22")
-        buf.write("\3\2\2\2\22\27\3\2\2\2\23\24\7\17\2\2\24\26\5\n\6\2\25")
-        buf.write("\23\3\2\2\2\26\31\3\2\2\2\27\25\3\2\2\2\27\30\3\2\2\2")
-        buf.write("\30\32\3\2\2\2\31\27\3\2\2\2\32\33\7\f\2\2\33\3\3\2\2")
-        buf.write("\2\34\37\7\b\2\2\35\36\7\r\2\2\36 \7\b\2\2\37\35\3\2\2")
-        buf.write("\2\37 \3\2\2\2 #\3\2\2\2!\"\7\16\2\2\"$\7\b\2\2#!\3\2")
-        buf.write("\2\2#$\3\2\2\2$\5\3\2\2\2%\'\7\34\2\2&%\3\2\2\2&\'\3\2")
-        buf.write("\2\2\'(\3\2\2\2()\5\4\3\2)\7\3\2\2\2*+\5\n\6\2+,\7\2\2")
-        buf.write("\3,\t\3\2\2\2-.\b\6\1\2./\7\7\2\2/\62\5\n\6\7\60\62\5")
-        buf.write("\f\7\2\61-\3\2\2\2\61\60\3\2\2\2\62>\3\2\2\2\63\64\f\6")
-        buf.write("\2\2\64\65\t\2\2\2\65=\5\n\6\7\66\67\f\5\2\2\678\t\3\2")
-        buf.write("\28=\5\n\6\69:\f\4\2\2:;\t\4\2\2;=\5\n\6\5<\63\3\2\2\2")
-        buf.write("<\66\3\2\2\2<9\3\2\2\2=@\3\2\2\2><\3\2\2\2>?\3\2\2\2?")
-        buf.write("\13\3\2\2\2@>\3\2\2\2AB\7\13\2\2BC\5\n\6\2CD\7\f\2\2D")
-        buf.write("L\3\2\2\2EL\t\5\2\2FL\5\2\2\2GL\t\6\2\2HL\5\6\4\2IL\7")
-        buf.write("\n\2\2JL\7\t\2\2KA\3\2\2\2KE\3\2\2\2KF\3\2\2\2KG\3\2\2")
-        buf.write("\2KH\3\2\2\2KI\3\2\2\2KJ\3\2\2\2L\r\3\2\2\2\13\21\27\37")
-        buf.write("#&\61<>K")
+        buf.write("\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\3\6\5\6")
+        buf.write("@\n\6\3\6\7\6C\n\6\f\6\16\6F\13\6\3\7\3\7\3\7\3\7\3\7")
+        buf.write("\3\7\3\7\3\7\3\7\3\7\5\7R\n\7\3\7\2\3\n\b\2\4\6\b\n\f")
+        buf.write("\2\7\3\2\34\35\3\2\32\33\3\2\22\31\3\2\3\4\3\2\5\6\2^")
+        buf.write("\2\16\3\2\2\2\4\34\3\2\2\2\6&\3\2\2\2\b*\3\2\2\2\n\61")
+        buf.write("\3\2\2\2\fQ\3\2\2\2\16\17\7\n\2\2\17\21\7\r\2\2\20\22")
+        buf.write("\5\n\6\2\21\20\3\2\2\2\21\22\3\2\2\2\22\27\3\2\2\2\23")
+        buf.write("\24\7\21\2\2\24\26\5\n\6\2\25\23\3\2\2\2\26\31\3\2\2\2")
+        buf.write("\27\25\3\2\2\2\27\30\3\2\2\2\30\32\3\2\2\2\31\27\3\2\2")
+        buf.write("\2\32\33\7\16\2\2\33\3\3\2\2\2\34\37\7\n\2\2\35\36\7\17")
+        buf.write("\2\2\36 \7\n\2\2\37\35\3\2\2\2\37 \3\2\2\2 #\3\2\2\2!")
+        buf.write("\"\7\20\2\2\"$\7\n\2\2#!\3\2\2\2#$\3\2\2\2$\5\3\2\2\2")
+        buf.write("%\'\7\36\2\2&%\3\2\2\2&\'\3\2\2\2\'(\3\2\2\2()\5\4\3\2")
+        buf.write(")\7\3\2\2\2*+\5\n\6\2+,\7\2\2\3,\t\3\2\2\2-.\b\6\1\2.")
+        buf.write("/\7\7\2\2/\62\5\n\6\b\60\62\5\f\7\2\61-\3\2\2\2\61\60")
+        buf.write("\3\2\2\2\62D\3\2\2\2\63\64\f\7\2\2\64\65\t\2\2\2\65C\5")
+        buf.write("\n\6\b\66\67\f\6\2\2\678\t\3\2\28C\5\n\6\79:\f\5\2\2:")
+        buf.write(";\t\4\2\2;C\5\n\6\6<=\f\4\2\2=?\7\b\2\2>@\7\7\2\2?>\3")
+        buf.write("\2\2\2?@\3\2\2\2@A\3\2\2\2AC\7\t\2\2B\63\3\2\2\2B\66\3")
+        buf.write("\2\2\2B9\3\2\2\2B<\3\2\2\2CF\3\2\2\2DB\3\2\2\2DE\3\2\2")
+        buf.write("\2E\13\3\2\2\2FD\3\2\2\2GH\7\r\2\2HI\5\n\6\2IJ\7\16\2")
+        buf.write("\2JR\3\2\2\2KR\t\5\2\2LR\5\2\2\2MR\t\6\2\2NR\5\6\4\2O")
+        buf.write("R\7\f\2\2PR\7\13\2\2QG\3\2\2\2QK\3\2\2\2QL\3\2\2\2QM\3")
+        buf.write("\2\2\2QN\3\2\2\2QO\3\2\2\2QP\3\2\2\2R\r\3\2\2\2\f\21\27")
+        buf.write("\37#&\61?BDQ")
         return buf.getvalue()
 
 
@@ -52,14 +54,15 @@ class TelParser ( Parser ):
 
     sharedContextCache = PredictionContextCache()
 
-    literalNames = [ "<INVALID>", "<INVALID>", "<INVALID>", "'true'", "'false'", 
-                     "'not'", "<INVALID>", "<INVALID>", "<INVALID>", "'('", 
-                     "')'", "'|'", "':'", "','", "'||'", "'&&'", "'=='", 
-                     "'!='", "'>'", "'<'", "'>='", "'<='", "'+'", "'-'", 
-                     "'*'", "'/'", "'?'" ]
+    literalNames = [ "<INVALID>", "<INVALID>", "<INVALID>", "<INVALID>", 
+                     "<INVALID>", "<INVALID>", "<INVALID>", "<INVALID>", 
+                     "<INVALID>", "<INVALID>", "<INVALID>", "'('", "')'", 
+                     "'|'", "':'", "','", "'||'", "'&&'", "'=='", "'!='", 
+                     "'>'", "'<'", "'>='", "'<='", "'+'", "'-'", "'*'", 
+                     "'/'", "'?'" ]
 
     symbolicNames = [ "<INVALID>", "INT", "REAL", "TRUE", "FALSE", "NOT", 
-                      "WORD", "STRING_CONSTANT", "SINGLE_QUOTED_ELEMENT", 
+                      "KW_IS", "KW_NULL", "WORD", "STRING_CONSTANT", "SINGLE_QUOTED_ELEMENT", 
                       "L_BRACKET", "R_BRACKET", "TAXON_NAMESPACE_DELIMITER", 
                       "TAXON_TAG_DELIMITER", "FN_PARAMETER_DELIMITER", "OR", 
                       "AND", "EQ", "NEQ", "GT", "LT", "GTEQ", "LTEQ", "PLUS", 
@@ -81,28 +84,30 @@ class TelParser ( Parser ):
     TRUE=3
     FALSE=4
     NOT=5
-    WORD=6
-    STRING_CONSTANT=7
-    SINGLE_QUOTED_ELEMENT=8
-    L_BRACKET=9
-    R_BRACKET=10
-    TAXON_NAMESPACE_DELIMITER=11
-    TAXON_TAG_DELIMITER=12
-    FN_PARAMETER_DELIMITER=13
-    OR=14
-    AND=15
-    EQ=16
-    NEQ=17
-    GT=18
-    LT=19
-    GTEQ=20
-    LTEQ=21
-    PLUS=22
-    MINUS=23
-    MULT=24
-    DIV=25
-    OPTIONAL_TAXON_OPERATOR=26
-    WS=27
+    KW_IS=6
+    KW_NULL=7
+    WORD=8
+    STRING_CONSTANT=9
+    SINGLE_QUOTED_ELEMENT=10
+    L_BRACKET=11
+    R_BRACKET=12
+    TAXON_NAMESPACE_DELIMITER=13
+    TAXON_TAG_DELIMITER=14
+    FN_PARAMETER_DELIMITER=15
+    OR=16
+    AND=17
+    EQ=18
+    NEQ=19
+    GT=20
+    LT=21
+    GTEQ=22
+    LTEQ=23
+    PLUS=24
+    MINUS=25
+    MULT=26
+    DIV=27
+    OPTIONAL_TAXON_OPERATOR=28
+    WS=29
 
     def __init__(self, input:TokenStream, output:TextIO = sys.stdout):
         super().__init__(input, output)
@@ -404,6 +409,37 @@ class TelParser ( Parser ):
             super().copyFrom(ctx)
 
 
+    class NullTestExprContext(ExprContext):
+
+        def __init__(self, parser, ctx:ParserRuleContext): # actually a TelParser.ExprContext
+            super().__init__(parser)
+            self.copyFrom(ctx)
+
+        def expr(self):
+            return self.getTypedRuleContext(TelParser.ExprContext,0)
+
+        def KW_IS(self):
+            return self.getToken(TelParser.KW_IS, 0)
+        def KW_NULL(self):
+            return self.getToken(TelParser.KW_NULL, 0)
+        def NOT(self):
+            return self.getToken(TelParser.NOT, 0)
+
+        def enterRule(self, listener:ParseTreeListener):
+            if hasattr( listener, "enterNullTestExpr" ):
+                listener.enterNullTestExpr(self)
+
+        def exitRule(self, listener:ParseTreeListener):
+            if hasattr( listener, "exitNullTestExpr" ):
+                listener.exitNullTestExpr(self)
+
+        def accept(self, visitor:ParseTreeVisitor):
+            if hasattr( visitor, "visitNullTestExpr" ):
+                return visitor.visitNullTestExpr(self)
+            else:
+                return visitor.visitChildren(self)
+
+
     class NotExprContext(ExprContext):
 
         def __init__(self, parser, ctx:ParserRuleContext): # actually a TelParser.ExprContext
@@ -589,7 +625,7 @@ class TelParser ( Parser ):
                 self.state = 44
                 self.match(TelParser.NOT)
                 self.state = 45
-                self.expr(5)
+                self.expr(6)
                 pass
             elif token in [TelParser.INT, TelParser.REAL, TelParser.TRUE, TelParser.FALSE, TelParser.WORD, TelParser.STRING_CONSTANT, TelParser.SINGLE_QUOTED_ELEMENT, TelParser.L_BRACKET, TelParser.OPTIONAL_TAXON_OPERATOR]:
                 localctx = TelParser.AtomExprContext(self, localctx)
@@ -602,24 +638,24 @@ class TelParser ( Parser ):
                 raise NoViableAltException(self)
 
             self._ctx.stop = self._input.LT(-1)
-            self.state = 60
+            self.state = 66
             self._errHandler.sync(self)
-            _alt = self._interp.adaptivePredict(self._input,7,self._ctx)
+            _alt = self._interp.adaptivePredict(self._input,8,self._ctx)
             while _alt!=2 and _alt!=ATN.INVALID_ALT_NUMBER:
                 if _alt==1:
                     if self._parseListeners is not None:
                         self.triggerExitRuleEvent()
                     _prevctx = localctx
-                    self.state = 58
+                    self.state = 64
                     self._errHandler.sync(self)
-                    la_ = self._interp.adaptivePredict(self._input,6,self._ctx)
+                    la_ = self._interp.adaptivePredict(self._input,7,self._ctx)
                     if la_ == 1:
                         localctx = TelParser.MultiplicationExprContext(self, TelParser.ExprContext(self, _parentctx, _parentState))
                         self.pushNewRecursionContext(localctx, _startState, self.RULE_expr)
                         self.state = 49
-                        if not self.precpred(self._ctx, 4):
+                        if not self.precpred(self._ctx, 5):
                             from antlr4.error.Errors import FailedPredicateException
-                            raise FailedPredicateException(self, "self.precpred(self._ctx, 4)")
+                            raise FailedPredicateException(self, "self.precpred(self._ctx, 5)")
                         self.state = 50
                         localctx.op = self._input.LT(1)
                         _la = self._input.LA(1)
@@ -629,16 +665,16 @@ class TelParser ( Parser ):
                             self._errHandler.reportMatch(self)
                             self.consume()
                         self.state = 51
-                        self.expr(5)
+                        self.expr(6)
                         pass
 
                     elif la_ == 2:
                         localctx = TelParser.AdditiveExprContext(self, TelParser.ExprContext(self, _parentctx, _parentState))
                         self.pushNewRecursionContext(localctx, _startState, self.RULE_expr)
                         self.state = 52
-                        if not self.precpred(self._ctx, 3):
+                        if not self.precpred(self._ctx, 4):
                             from antlr4.error.Errors import FailedPredicateException
-                            raise FailedPredicateException(self, "self.precpred(self._ctx, 3)")
+                            raise FailedPredicateException(self, "self.precpred(self._ctx, 4)")
                         self.state = 53
                         localctx.op = self._input.LT(1)
                         _la = self._input.LA(1)
@@ -648,16 +684,16 @@ class TelParser ( Parser ):
                             self._errHandler.reportMatch(self)
                             self.consume()
                         self.state = 54
-                        self.expr(4)
+                        self.expr(5)
                         pass
 
                     elif la_ == 3:
                         localctx = TelParser.LogicalExprContext(self, TelParser.ExprContext(self, _parentctx, _parentState))
                         self.pushNewRecursionContext(localctx, _startState, self.RULE_expr)
                         self.state = 55
-                        if not self.precpred(self._ctx, 2):
+                        if not self.precpred(self._ctx, 3):
                             from antlr4.error.Errors import FailedPredicateException
-                            raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
+                            raise FailedPredicateException(self, "self.precpred(self._ctx, 3)")
                         self.state = 56
                         localctx.op = self._input.LT(1)
                         _la = self._input.LA(1)
@@ -667,13 +703,34 @@ class TelParser ( Parser ):
                             self._errHandler.reportMatch(self)
                             self.consume()
                         self.state = 57
-                        self.expr(3)
+                        self.expr(4)
+                        pass
+
+                    elif la_ == 4:
+                        localctx = TelParser.NullTestExprContext(self, TelParser.ExprContext(self, _parentctx, _parentState))
+                        self.pushNewRecursionContext(localctx, _startState, self.RULE_expr)
+                        self.state = 58
+                        if not self.precpred(self._ctx, 2):
+                            from antlr4.error.Errors import FailedPredicateException
+                            raise FailedPredicateException(self, "self.precpred(self._ctx, 2)")
+                        self.state = 59
+                        self.match(TelParser.KW_IS)
+                        self.state = 61
+                        self._errHandler.sync(self)
+                        _la = self._input.LA(1)
+                        if _la==TelParser.NOT:
+                            self.state = 60
+                            self.match(TelParser.NOT)
+
+
+                        self.state = 63
+                        self.match(TelParser.KW_NULL)
                         pass
 
              
-                self.state = 62
+                self.state = 68
                 self._errHandler.sync(self)
-                _alt = self._interp.adaptivePredict(self._input,7,self._ctx)
+                _alt = self._interp.adaptivePredict(self._input,8,self._ctx)
 
         except RecognitionException as re:
             localctx.exception = re
@@ -886,24 +943,24 @@ class TelParser ( Parser ):
         self.enterRule(localctx, 10, self.RULE_atom)
         self._la = 0 # Token type
         try:
-            self.state = 73
+            self.state = 79
             self._errHandler.sync(self)
-            la_ = self._interp.adaptivePredict(self._input,8,self._ctx)
+            la_ = self._interp.adaptivePredict(self._input,9,self._ctx)
             if la_ == 1:
                 localctx = TelParser.BracketExprContext(self, localctx)
                 self.enterOuterAlt(localctx, 1)
-                self.state = 63
+                self.state = 69
                 self.match(TelParser.L_BRACKET)
-                self.state = 64
+                self.state = 70
                 self.expr(0)
-                self.state = 65
+                self.state = 71
                 self.match(TelParser.R_BRACKET)
                 pass
 
             elif la_ == 2:
                 localctx = TelParser.NumberAtomContext(self, localctx)
                 self.enterOuterAlt(localctx, 2)
-                self.state = 67
+                self.state = 73
                 _la = self._input.LA(1)
                 if not(_la==TelParser.INT or _la==TelParser.REAL):
                     self._errHandler.recoverInline(self)
@@ -915,14 +972,14 @@ class TelParser ( Parser ):
             elif la_ == 3:
                 localctx = TelParser.FnExprContext(self, localctx)
                 self.enterOuterAlt(localctx, 3)
-                self.state = 68
+                self.state = 74
                 self.fn()
                 pass
 
             elif la_ == 4:
                 localctx = TelParser.BooleanAtomContext(self, localctx)
                 self.enterOuterAlt(localctx, 4)
-                self.state = 69
+                self.state = 75
                 _la = self._input.LA(1)
                 if not(_la==TelParser.TRUE or _la==TelParser.FALSE):
                     self._errHandler.recoverInline(self)
@@ -934,21 +991,21 @@ class TelParser ( Parser ):
             elif la_ == 5:
                 localctx = TelParser.TaxonSlugAtomContext(self, localctx)
                 self.enterOuterAlt(localctx, 5)
-                self.state = 70
+                self.state = 76
                 self.taxon_expr()
                 pass
 
             elif la_ == 6:
                 localctx = TelParser.SingleQuotedAtomContext(self, localctx)
                 self.enterOuterAlt(localctx, 6)
-                self.state = 71
+                self.state = 77
                 self.match(TelParser.SINGLE_QUOTED_ELEMENT)
                 pass
 
             elif la_ == 7:
                 localctx = TelParser.StringConstantAtomContext(self, localctx)
                 self.enterOuterAlt(localctx, 7)
-                self.state = 72
+                self.state = 78
                 self.match(TelParser.STRING_CONSTANT)
                 pass
 
@@ -975,14 +1032,18 @@ class TelParser ( Parser ):
 
     def expr_sempred(self, localctx:ExprContext, predIndex:int):
             if predIndex == 0:
-                return self.precpred(self._ctx, 4)
+                return self.precpred(self._ctx, 5)
          
 
             if predIndex == 1:
-                return self.precpred(self._ctx, 3)
+                return self.precpred(self._ctx, 4)
          
 
             if predIndex == 2:
+                return self.precpred(self._ctx, 3)
+         
+
+            if predIndex == 3:
                 return self.precpred(self._ctx, 2)